# gpu concat optimize