before:
```
Benchmarking tfl/1024/mxnet-ngraph-Function_1-MakeForwardFunction-pre-optimized-fprop.json, GPU backend, 10 iterations.
compile time: 12,709ms
703ms per iteration

---- Aggregate times per op type ----
Concat        426,170us
Softmax       122,471us
Broadcast      52,662us
Add            25,070us
Slice          19,250us
Dot            16,340us
Multiply        5,766us
Divide          3,535us
Exp             2,889us
Negative        2,881us
Reshape         2,617us
Result          2,258us
Tanh            2,007us

---- Aggregate times per op type/shape/count ----
Concat {1024, 60, 200} 1      426,170us
Softmax {61440, 2000} 1       122,471us
Broadcast {1024, 800} 240      26,035us
Slice {1024, 200} 480          16,315us
Broadcast {61440, 2000} 1      15,710us
Add {1024, 800} 360            12,956us
Dot {1024, 800} 240            12,401us
Broadcast {1024, 200} 364      10,917us
Add {1024, 200} 480             7,678us
Multiply {1024, 200} 360        5,766us
Add {61440, 2000} 1             4,436us
Dot {61440, 2000} 1             3,939us
Divide {1024, 200} 360          3,535us
Slice {1024, 1, 200} 60         2,935us
Exp {1024, 200} 360             2,889us
Negative {1024, 200} 360        2,881us
Result {61440, 2000} 1          2,258us
Tanh {1024, 200} 240            2,007us
Reshape {200, 800} 240          1,449us
Reshape {1024, 200} 60            472us
Reshape {1024, 1, 200} 60         446us
Reshape {61440, 200} 1            236us
Reshape {200, 2000} 1              14us
```

after:
```
Benchmarking tfl/1024/mxnet-ngraph-Function_1-MakeForwardFunction-pre-optimized-fprop.json, GPU backend, 10 iterations.
compile time: 11,082ms
287.3ms per iteration

---- Aggregate times per op type ----
Softmax       117,133us
Broadcast      51,413us
Add            25,077us
Slice          18,944us
Concat         17,634us
Dot            16,195us
Multiply        5,764us
Divide          3,539us
Exp             2,886us
Negative        2,883us
Reshape         2,577us
Result          2,258us
Tanh            1,940us

---- Aggregate times per op type/shape/count ----
Softmax {61440, 2000} 1       117,133us
Broadcast {1024, 800} 240      25,490us
Concat {1024, 60, 200} 1       17,634us
Slice {1024, 200} 480          16,079us
Broadcast {61440, 2000} 1      15,365us
Add {1024, 800} 360            12,951us
Dot {1024, 800} 240            12,220us
Broadcast {1024, 200} 364      10,558us
Add {1024, 200} 480             7,692us
Multiply {1024, 200} 360        5,764us
Add {61440, 2000} 1             4,434us
Dot {61440, 2000} 1             3,975us
Divide {1024, 200} 360          3,539us
Exp {1024, 200} 360             2,886us
Negative {1024, 200} 360        2,883us
Slice {1024, 1, 200} 60         2,865us
Result {61440, 2000} 1          2,258us
Tanh {1024, 200} 240            1,940us
Reshape {200, 800} 240          1,448us
Reshape {1024, 200} 60            449us
Reshape {1024, 1, 200} 60         435us
Reshape {61440, 200} 1            231us
Reshape {200, 2000} 1              14us
```